% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PicMin_functions.R
\name{EmpiricalPs}
\alias{EmpiricalPs}
\title{Calculate empirical p-values from a vector of summary statistics}
\usage{
EmpiricalPs(vector_of_values, large_i_small_p = FALSE)
}
\arguments{
\item{vector_of_values}{A set of summary statistics that you want to convert to empirical p-values}

\item{large_i_small_p}{Do you want large values to have small p-values (e.g. Fst)?}
}
\description{
Calculate empirical p-values from a vector of summary statistics
}
