% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PicMin_functions.R
\name{orderStatsPValues}
\alias{orderStatsPValues}
\title{orderStatsPValues}
\usage{
orderStatsPValues(p_list)
}
\arguments{
\item{p_list}{the vector of p-values from your genome scans}
}
\description{
orderStatsPValues
}
