% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PicMin_functions.R
\name{PicMin}
\alias{PicMin}
\title{Generate data under the null hypothesis}
\usage{
PicMin(pList, correlationMatrix, numReps = 1e+05)
}
\arguments{
\item{pList}{the vector of p-values from your genome scans}

\item{correlationMatrix}{the correlation matrix under the null hypothesis}

\item{numReps}{the number of replicate draws to perform when building the empirical distributing for calculating the Tippett p-value}
}
\description{
Generate data under the null hypothesis
}
